#260: Made default Options.init to read Info.plist
