+++ AtlasSDK/AtlasSDK/OptionsInfoPlist.swift
@@ -4,27 +4,33 @@
 
 import Foundation
 
-public extension Options {
+extension Options {
 
-    public init(bundle: NSBundle = NSBundle.mainBundle()) {
-        self.init(clientId: bundle.string(.clientId),
-            salesChannel: bundle.string(.salesChannel),
-            useSandbox: bundle.bool(.useSandbox),
-            countryCode: bundle.string(.countryCode),
-            interfaceLanguage: bundle.string(.interfaceLanguage)
-        )
+    enum InfoKey: String {
+
+        case useSandbox = "ATLASSDK_USE_SANDBOX"
+        case clientId = "ATLASSDK_CLIENT_ID"
+        @available( *, deprecated, message = "To be removed when backend returns correct sales channel")
+        case salesChannel = "ATLASSDK_SALES_CHANNEL"
+        case interfaceLanguage = "ATLASSDK_INTERFACE_LANGUAGE"
+        case countryCode = "ATLASSDK_COUNTRY_CODE"
+
     }
 
 }
 
 extension NSBundle {
 
-    private func string(key: InfoKey, defaultValue: String = "") -> String {
-        return self.objectForInfoDictionaryKey(key.rawValue) as? String ?? defaultValue
+    func string(key: Options.InfoKey, defaultValue: String = "") -> String {
+        return self.infoObject(forKey: key) ?? defaultValue
     }
 
-    private func bool(key: InfoKey, defaultValue: Bool = false) -> Bool {
-        return self.objectForInfoDictionaryKey(key.rawValue) as? Bool ?? defaultValue
+    func bool(key: Options.InfoKey, defaultValue: Bool = false) -> Bool {
+        return self.infoObject(forKey: key) ?? defaultValue
+    }
+
+    func infoObject<T>(forKey key: Options.InfoKey) -> T? {
+        return self.objectForInfoDictionaryKey(key.rawValue) as? T
     }
 
 }
